Login box created

+++ tool-library-java-backend/src/main/java/toollibrary/api/ToolController.java
@@ -1,7 +1,10 @@
 package toollibrary.api;
 
+import toollibrary.component.MemberComponent;
 import toollibrary.dao.ToolDao;
+import toollibrary.exception.ForbiddenException;
 import toollibrary.exception.ResourceNotFoundException;
+import toollibrary.model.Member;
 import toollibrary.model.Tool;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -16,15 +19,23 @@ public class ToolController {
     @Autowired
     private ToolDao toolDao;
 
+    @Autowired
+    private MemberComponent memberComponent;
+
     // get all tools
     @GetMapping
-    public List<Tool> getAllTools() { // we could use our own data structure here
+    public List<Tool> getAllTools() {
         return toolDao.findAll();
     }
 
-    @PostMapping
-    public Tool createTool(@RequestBody Tool tool) {
-        return toolDao.save(tool);
+    // create a tool
+    @PostMapping("/{user}")
+    public ResponseEntity<Tool> createTool(@PathVariable String user, @RequestBody Tool tool) {
+        Member member = memberComponent.checkCredentials(user);
+
+        tool.setOwner(member);
+        
+        return ResponseEntity.ok(toolDao.save(tool));
     }
 
     // method to return a single tool by id
@@ -33,16 +44,25 @@ public class ToolController {
     @GetMapping("/{id}")
     public ResponseEntity<Tool> getToolById(@PathVariable long id) {
         Tool tool = toolDao.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Tool not found with id: " + id));
+            .orElseThrow(() -> new ResourceNotFoundException());
+        
         return ResponseEntity.ok(tool);
     }
 
     // method to update a tool by id
     // this one requires a tool object to be passed in the body of the request
-    @PutMapping("/{id}")
-    public ResponseEntity<Tool> updateToolById(@PathVariable long id,@RequestBody Tool tool) {
+    @PutMapping("/{id}/{user}")
+    public ResponseEntity<Tool> updateToolById(
+            @PathVariable long id, @PathVariable String user, @RequestBody Tool tool) {
         Tool toolToUpdate = toolDao.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Tool not found with id: " + id));
+            .orElseThrow(() -> new ResourceNotFoundException());
+
+        Member member = memberComponent.checkCredentials(user);
+        Member owner = toolToUpdate.getOwner();
+
+        if (owner == null || owner.getId() != member.getId()) {
+            throw new ForbiddenException();
+        }
 
         toolToUpdate.setToolName(tool.getToolName());
         toolToUpdate.setToolDescription(tool.getToolDescription());
@@ -54,10 +74,18 @@ public class ToolController {
     }
 
     // method to delete a tool by id
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Tool> deleteToolById(@PathVariable long id) {
+    @DeleteMapping("/{id}/{user}")
+    public ResponseEntity<Tool> deleteToolById(@PathVariable long id, @PathVariable String user) {
         Tool toolToDelete = toolDao.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Tool not found with id: " + id));
+            .orElseThrow(() -> new ResourceNotFoundException());
+
+        Member member = memberComponent.checkCredentials(user);
+        Member owner = toolToDelete.getOwner();
+
+        if (owner == null || owner.getId() != member.getId()) {
+            throw new ForbiddenException();
+        }
+
         toolDao.delete(toolToDelete);
         return ResponseEntity.ok().build();
     }
